requirements.txt
appnope==0.1.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
Django==1.11.20
django-filter==2.0.0
djangorestframework==3.9.2
enum34==1.1.6
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
kiwisolver==1.1.0
lxml==4.3.3
Markdown==3.1
matplotlib==2.2.4
mysqlclient==1.4.2.post1
numpy==1.16.3
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
soupsieve==1.9.1
subprocess32==3.5.3
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
